changed

+++ src/Flowers/payment/CreditCardPaymentStrategy.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 
 public class CreditCardPaymentStrategy implements IPayment {
-    public boolean pay(double price){
+    public boolean pay(int price){
         Random rndGenerator = new Random();
         boolean [] newLst = {true, false};
         return newLst[rndGenerator.nextInt(1)];
